Add files via upload
Classification of cars. The dataset is customised one and images could be downloaded from Google Images..The classification is done using FastAi-Libraries.The model used here is ResNet34  with pre-trained weights. The Images are taken from google(Randomly 100-black,100-toy,100-Vintage). This Could be applied to any set of classification for images